+++ test/oap3/resources/OAR043/ok-example.js
@@ -1,0 +1,39 @@
+module.exports = {
+  "openapi": "3.0.0",
+  "info": {
+    "version": "1.0.0",
+    "title": "OAR043 ok"
+  },
+  "paths": {
+    "/orders": {
+      "get": {
+        "parameters": [
+          {
+            "name": "status",
+            "in": "query",
+            "required": false,
+            "schema": { "type": "string" }
+          }
+        ],
+        "responses": {
+          "200": { "description": "OK" }
+        }
+      }
+    },
+    "/orders/{id}": {
+      "get": {
+        "parameters": [
+          {
+            "in": "path",
+            "name": "id",
+            "required": true,
+            "schema": { "type": "integer" }
+          }
+        ],
+        "responses": {
+          "200": { "description": "OK" }
+        }
+      }
+    }
+  }
+};